feat: 초대장 상세 페이지 api 로딩, 에러 처리
