+++ apps/tuk-web/src/shared/lib/api/rest/index.ts
@@ -6,27 +6,27 @@ import { getBridgeSender } from '@/shared/lib/app-bridge/bridgeSender';
 export { RestAPI } from './rest';
 export type { RestAPIConfig, RestAPIProtocol } from './types';
 
-// const defaultJsonInstance = (baseURL: string) =>
-//   new RestAPIInstance(baseURL, {
-//     withCredentials: false,
-//     authHeader: () => {
-//       if (typeof window === 'undefined') return null;
-//       const token = sessionStorage.getItem('accessToken');
-//       return token ? { Authorization: `Bearer ${token}` } : null;
-//     },
-//     bridgeSend: msg => getBridgeSender()?.(msg),
-//   });
-
 const defaultJsonInstance = (baseURL: string) =>
   new RestAPIInstance(baseURL, {
     withCredentials: false,
     authHeader: () => {
       if (typeof window === 'undefined') return null;
-      const token =
-        'eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIzMiIsIm1lbWJlcklkIjoiMzIiLCJ0b2tlblR5cGUiOiJBQ0NFU1MiLCJpYXQiOjE3NTUwMDM4NTQsImV4cCI6MTc1NzU5NTg1NH0.sXHpNsim8-nIUdCjnAcOIZGsg2gUo8rnpkBmN-z5dKU';
-      return token ? { Authorization: `Bearer ${token.trim()}` } : null;
+      const token = sessionStorage.getItem('accessToken');
+      return token ? { Authorization: `Bearer ${token}` } : null;
     },
+    bridgeSend: msg => getBridgeSender()?.(msg),
   });
+
+// const defaultJsonInstance = (baseURL: string) =>
+//   new RestAPIInstance(baseURL, {
+//     withCredentials: false,
+//     authHeader: () => {
+//       if (typeof window === 'undefined') return null;
+//       const token =
+//         'eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIzMiIsIm1lbWJlcklkIjoiMzIiLCJ0b2tlblR5cGUiOiJBQ0NFU1MiLCJpYXQiOjE3NTUwMDM4NTQsImV4cCI6MTc1NzU5NTg1NH0.sXHpNsim8-nIUdCjnAcOIZGsg2gUo8rnpkBmN-z5dKU';
+//       return token ? { Authorization: `Bearer ${token.trim()}` } : null;
+//     },
+//   });
 
 const formInstance = (baseURL: string) =>
   new RestAPIInstance(baseURL, {
